+++ src/jmetal/metaheuristics/moead/MOEAD_SAS_main.java
@@ -186,9 +186,9 @@ public class MOEAD_SAS_main extends SASAlgorithmAdaptor{
 
 		// Algorithm parameters
 		int popsize = 600;
-		int generations = 5;
+		int factor = 5;
 		algorithm.setInputParameter("populationSize", popsize);
-		algorithm.setInputParameter("maxEvaluations", popsize * generations);
+		algorithm.setInputParameter("maxEvaluations", popsize * factor);
 		
 		algorithm.setInputParameter("dataDirectory", "weight");
 
@@ -258,10 +258,13 @@ public class MOEAD_SAS_main extends SASAlgorithmAdaptor{
 		
 		String str = "data/" +problem.getName()
 		+ "M" + problem.getNumberOfObjectives() + "/SAS";
+		if(SAS.isTest) 
 		Utils.deleteFolder(new File(str+ "/knee_results.dat"));
+		
 		SolutionSet set = new SolutionSet(1);
 		set.add(kneeIndividual);
 		
+		if(SAS.isTest) 
 		set.printObjectivesToFile(str + "/knee_results.dat");
 		
 		return kneeIndividual;
